VFS/LibC: Create ftruncate function and corresponding syscall and libc implementation
Previously this function was only used for shared memory region created
by shm_open because I was lazy. Now exists for all files.

+++ userland/libc/include/unistd.h
@@ -10,7 +10,7 @@ extern int opterr, optind, optopt;
 extern char *optarg;
 
 int close(int fildes);
-int ftruncate(int fildes, uint64_t length);
+int ftruncate(int fildes, size_t length);
 int execv(char *path, char **argv);
 int pipe(int fd[2]);
 int dup2(int org_fd, int new_fd);
